// src/app/shared/interfaces/isea-farer.ts
export interface ISeaFarer {
    Id: number
    KinName: any
    KinPhone: any
    KinEmail: any
    PassPortIssueDate: any
    EmpId: number
    VisaSponsorId: number
    Remarks: string
    VisaIssueDate: any
    VisaExpiryDate: any
    EmployeeName: string
    SponsorName: string
    EmployeeCode: string
    JobName: string
    Nationality: string
    BirthDate: string
    Phone: string
    Mobile: string
    Email: string
    NationalId: string
    EmploymentDate: string
    InsuranceDate: any
    PassportExpireDate: any
    PassportNumber: string
    Age: string
    EmployeeNameAR: string
    EmployeeNameEN: string
    JobNameEN: string
    JobNameAR: string
    Salary: number
    Address: string
    Gender: string
    Religion: string
    InsuranceNumber: any
    Area: any
    MilitaryServiceExpireDate: any
    SyndicateCardNumber: any
    SyndicateExpirationDate: any
    BusinessPhone: any
    BirthPlace: string
    NationalIdExpiryDate: any
    Company: any
    PostCategoryLevel: any
    Tier: string
    DrivingLicenseNumber: any
    WorkPermit: any
    PermitNumber: any
    BloodType: any
    WorkLocation: any
    PermitExpirationDate: any
    Notes: any
    MedicalInsuranceAmount: any
    InsuranceTitle: any
    InsuranceOffice: any
    MedicalInsuranceNo: any
    MedicalInsuranceDate: any
    GraduationYear: any
    PostGraduateDegree: any
    PostGraduateUniversity: any
    PostGraduateSpecialty: any
    PostGraduateGrade: any
    PropationEndDate: string
    EmpManager: string
    NameOfSpouse: string
    NoOfChildren: number
    BodyWeight: number
    Height: number
    VisaUAEIdNO: string
    NearestAirport: string
    ResidenceNumber: string
    SkypeID: string
    PermanentAddressHomeCountry: string
    ContactNumberHomeCountry: string
    ContactNameAndNumberDuringEmergenciesUAE: string
    ContactNameAndNumberDuringEmergenciesHome: string
    SeamanBookNO: string
    SeamanIssueDate: any
    SeamanExpiryDate: any
    CicpaNO: string
    CicpaIssueDate: any
    CicpaExpiryDate: any
    Declaration: string
    SignedOffFromAShipDueToMedicalReason: boolean
    SignedOffFromAShipDueToMedicalReasonComment: string
    UndergoneAnyMdicalOperation: any
    UndergoneAnyMdicalOperationComment: any
    DoctorConsultation: boolean
    DoctorConsultationComment: string
    HealthOrDisabilityProblem: boolean
    HealthOrDisabilityProblemComment: string
    InquiryOrInvolvedMaritimeAccident: boolean
    InquiryOrInvolvedMaritimeAccidentComment: string
    LicenseSuspendedOrRevoked: boolean
    LicenseSuspendedOrRevokedComment: string
    ConfirmationInsuranceDate: any
    MaritalStatus: string
    Status: number
    Basic: any
    Advanced: any
    FullLicence: any
    Maintenance: any
    isActive?: boolean;
}